requirements.txt
# ====================================
# DEPENDENCIAS DEL PROYECTO
# ====================================
# Instalar con: pip install -r requirements.txt

# Selenium para automatización web
selenium>=4.0.0

# Pandas para procesamiento de datos
pandas>=2.0.0

# Base de datos
sqlalchemy>=2.0.0
pyodbc>=4.0.0

# Archivos Excel
openpyxl>=3.1.0

# Variables de entorno
python-dotenv>=1.0.0

# Opcional: para testing
# pytest>=7.0.0
